Fix Android binding native packaging
